[refactor] button components
