+++ renderer/components/buttons/EndBtn.tsx
@@ -2,7 +2,7 @@ import { Sprite } from '@inlet/react-pixi'
 import { InteractionEvent } from 'pixi.js'
 import { useState } from 'react'
 
-import { containsPoint } from '../utils/pixi_api'
+import { containsPoint } from '../../utils/pixi_api'
 
 interface Props {
   handleStopClick: (event: InteractionEvent) => void
